Major Changes
Decorated the test map with functional obstacles.
Made the Player a prefab
Added public variables to the players toilet paper script to edit during design process
added public variables to the enemy spawner script
Changed the "Walls" layer to "Obstacles" to remove any confusion as to what layer is used for collision detection.